requirements.txt
# Primetrade Assignment - Data Science Pipeline
# Python 3.9+

# Data & config
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0

# Google Drive download (for datasets)
gdown>=4.6.0
requests>=2.28.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# ML & imbalanced handling
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0

# SHAP for model explanations
shap>=0.41.0

# Dashboard
streamlit>=1.25.0
